chore: rename countryCode.json file
Pluralize
