+++ src/components/invoices/InvoiceCustomerInfos.tsx
@@ -6,7 +6,7 @@ import { generatePath, Link } from 'react-router-dom'
 import { Typography } from '~/components/designSystem'
 import { CountryCode, InvoiceForInvoiceInfosFragment } from '~/generated/graphql'
 import { CUSTOMER_DETAILS_ROUTE } from '~/core/router'
-import CountryCodes from '~/public/countryCode.json'
+import CountryCodes from '~/public/countryCodes.json'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 import { theme } from '~/styles'
 import { formatDateToTZ } from '~/core/timezone'
